+++ docs/en/changes/changes.md
@@ -36,6 +36,8 @@
 * chore: add `toString` implementation for `StorageID`.
 * chore: add a warning log when connecting to ES takes too long.
 * Fix the query time range in the metadata API.
+* OAP gRPC-Client support `Health Check`.
+* [Break Change] `Health Check` make response 1 represents healthy, 0 represents unhealthy.
 
 #### UI